refactor(server): Split graphql.rs into smaller modules
Split the 2112-line graphql.rs file into logical modules:
- types.rs (42 lines): Request/response types
- where_clause.rs (704 lines): WHERE clause parsing and SQL generation
- introspection.rs (208 lines): Schema introspection support
- parser.rs (730 lines): Query/mutation parsing
- relationships.rs (433 lines): Nested query handling
- mod.rs (77 lines): Module exports

All 46 existing tests pass. No public API changes.

Closes #4320
